Update 04.addressing-clinical-questions.md
Trying to incorporate Pete's comment (have not found a paper to cite so i put in further work is needed): "A great recent example of this is COVID.  During the NYC and Wuhan peak of epidemic radiologists called COVID with high sensitivity (high prevalence).  Now, the prevalence has descended into the low single digits,  the detection rate has fallen off too.  This is not just  differences in perception but also an unwillingness to call something  as positive  recognizing its low prevalence (and massive implications for everyone involved)."